+++ src/Scaffolder/src/Config/ScaffolderConfig.php
@@ -30,9 +30,22 @@ class ScaffolderConfig extends InjectableConfig
         return $this->config['header'];
     }
 
+    /**
+     * @deprecated since v3.8.0.
+     */
     public function baseDirectory(): string
     {
         return $this->config['directory'];
+    }
+
+    /**
+     * @param non-empty-string $element
+     */
+    public function declarationDirectory(string $element): string
+    {
+        $declaration = $this->getDeclaration($element);
+
+        return !empty($declaration['directory']) ? $declaration['directory'] : $this->config['directory'];
     }
 
     /**
@@ -79,7 +92,7 @@ class ScaffolderConfig extends InjectableConfig
         $elementNamespace = \substr($elementNamespace, \strlen($this->baseNamespace($element)));
 
         return $this->joinPathChunks([
-            $this->baseDirectory(),
+            $this->declarationDirectory($element),
             \str_replace('\\', '/', $elementNamespace),
             $this->className($element, $name) . '.php',
         ], '/');